more gpio.  almost there
