+++ src/pirri/pirriGpioPin.go
@@ -1,5 +1,12 @@
 package main
 
+import (
+	"fmt"
+	"time"
+
+	"github.com/stianeikeland/go-rpio"
+)
+
 //GpioPin - describes a Raspberry Pi GPIO pin
 type GpioPin struct {
 	ID     int `sql:"AUTO_INCREMENT" gorm:"primary_key"`
@@ -8,53 +15,63 @@ type GpioPin struct {
 	Common bool `sql:"DEFAULT:false" gorm:"not null"`
 }
 
-func getGpioByPin() {
-	gpio := db.Where("GPIO = ?", 4).Find(&GpioPin{}).Order("GPIO DESC")
-	jsonifySQLResults(gpio)
+func setCommonWire() {
+	var gpio GpioPin
+	db.Where("common = ?", true).Limit(1).Find(&gpio)
+	COMMONWIRE = gpio.GPIO
 }
 
-// func (g *GpioPin) gpioActivator(gpio int, state bool, seconds int) {
-//  if SETTINGS.SimulateGpioActivity {
-//      gpioSimulation(gpio, state, seconds)
-//  } else {
-//      gpioActivate(gpio, state, seconds)
-//  }
-// }
+func gpioActivator(gpio int, state bool, seconds int) {
+	if SETTINGS.Debug.SimulateGPIO {
+		gpioSimulation(gpio, state, seconds)
+	} else {
+		gpioActivate(gpio, state, seconds)
+	}
+}
 
-// func (g *GpioPin) gpioSimulation(gpio int, state bool, seconds int) {
-//  fmt.Println("GPIO Simulation starting:",
-//      "\nTime:", time.Now(),
-//      "\nGPIO:", gpio,
-//      "\nDesired State:", state,
-//      "\nDuration (seconds):", seconds)
-//  fmt.Println("Active!", time.Now())
-//  for seconds > 0 {
-//      time.Sleep(time.Second)
-//      seconds--
-//  }
-//  fmt.Println("Deactivated!", time.Now())
-// }
+func gpioSimulation(gpio int, state bool, seconds int) {
+	fmt.Println("GPIO Simulation starting:",
+		"\nTime:", time.Now(),
+		"\nGPIO:", gpio,
+		"\nDesired State:", state,
+		"\nDuration (seconds):", seconds)
+	fmt.Println("Active!", time.Now())
+	for seconds > 0 && !RUNSTATUS.Cancel {
+		time.Sleep(time.Second)
+		seconds--
+	}
+	fmt.Println("Deactivated!", time.Now())
+}
 
-// func (g *GpioPin) gpioActivate(gpio int, state bool, seconds int) {
-//  pin := rpio.Pin(gpio)
-//  defer rpio.Close()
+func gpioActivate(gpio int, state bool, seconds int) {
+	rpio.Open()
+	defer rpio.Close()
+	pin := rpio.Pin(gpio)
+	common := rpio.Pin(COMMONWIRE)
+	pin.Output()
+	common.Output()
 
-//  // activate gpio
-//  if state {
-//      pin.High()
-//  } else {
-//      pin.Low()
-//  }
+	// activate gpio
+	if state {
+		common.High()
+		pin.High()
+	} else {
+		common.Low()
+		pin.Low()
+	}
 
-//  // start countdown
-//  for seconds > 0 {
-//      time.Sleep(time.Duration(1) * time.Second)
-//      seconds--
-//  }
+	// start countdown
+	for seconds > 0 && RUNSTATUS.Cancel {
+		time.Sleep(time.Duration(1) * time.Second)
+		seconds--
+	}
 
-//  if !state {
-//      pin.High()
-//  } else {
-//      pin.Low()
-//  }
-// }
+	// undo what we just did
+	if !state {
+		common.Low()
+		pin.High()
+	} else {
+		common.Low()
+		pin.Low()
+	}
+}
